done change language

+++ delete.js
@@ -1,5 +1,5 @@
 import {deleteFromDictionary} from './dictionary.js';
-import {warningInput, buttonGo} from './index.js';
+import {warningInput} from './index.js';
 
 
 let str = `
@@ -10,10 +10,18 @@ let str = `
 export function deleteWord () {
     main.insertAdjacentHTML ('beforeend', str );
     deleteButton.addEventListener ('click', deleteWorldHelp);
-    buttonGo (deleteWorldHelp)
+    document.addEventListener ('keydown', deleteWorldHelp);
+
+    document.addEventListener ('click', function cbf2 (event) {
+        if (event.target.textContent == 'Kraken Translate' || event.target.textContent == 'Home') {
+            document.removeEventListener ('click', cbf2);
+            document.removeEventListener ('keydown', deleteWorldHelp);
+        }  
+    })
 }
 
 function deleteWorldHelp () {
+    if (event.keyCode !== 13 && event.target.textContent != 'Delete world') return;
     if (!deleteFromDictionary (englishDelete.value.toLowerCase () )) {
         warningInput('Nope', 'red', englishDelete)
     } else {
